fastapi/requirements.txt
requests
langchain
openai
xmltodict
langchain-community
python-dotenv
langchain-openai
serpapi
google-search-results
google-serp-api
fastapi[standard]
snowflake-connector-python